+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 
 - Added http proxy for change cookie
+- Added settings: "proxy_port"
 
 ## 0.1.0